+++ src/index.ts
@@ -1,4 +1,4 @@
 import * as apollo from "./apollo"
 export { apollo }
-export { default as graphql } from "./graphql"
+export { default as query } from "./query"
 export { default as ApolloProp } from "./ApolloProp"