+++ CHANGELOG.md
@@ -7,8 +7,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 
 ## [Unreleased]
+### Changed
+- Replace `graphql()` with `query()` for flexibility.
 
 ## [0.1.0] - 2018-03-28
+### Added
 - Initial release.
 
 [Unreleased]: https://github.com/yuku-t/thyperapp-apolloextcomplete/compare/v0.1.0...HEAD